feat: Add GetEeclassBulletinContent usecase
This commit introduces the `GetEeclassBulletinContent` use case,
which retrieves the content of an EEClass bulletin.

It depends on the `EeclassRepository` and defines a `call` method
that takes `GetBulletinParams` as input and returns either a `Failure`
or an `EeclassBulletin` entity.

The `GetBulletinParams` class is also added to encapsulate the
necessary parameters for the use case.