feat: Implement usecase for getting assignment list
This commit introduces a new use case, GetEeclassCourseAssignmentList,
for retrieving a list of assignments for a given course. It also
includes the GetAssignmentListParams class for passing parameters to the
use case.

The use case depends on the EeclassRepository to fetch the data and
returns either a list of EeclassAssignmentInfo or a Failure.