feat: Implement GetEeclassCourseList use case
This commit implements the GetEeclassCourseList use case, which
retrieves a list of Eeclass courses.

The use case takes GetCourseListParams as input, which includes the
semester ID. It then calls the getCourseList method of the
EeclassRepository to fetch the course list. The result is returned as an
Either type, indicating either a successful retrieval of the course list
or a Failure.